refactor(network): split quic.rs (1578 LOC) into frame / stream modules
quic.rs は v1.0 PR レビューで oversized grab-bag として flag された唯一の
source file。 behavior-preserving な module 再編で coherent に分割する。

- frame.rs: typed-frame wire I/O (read/write_frame, read/write_typed_frame),
  FRAME_TYPE_* / CHANNEL_ACK_METHOD 定数, write_channel_ack
- stream.rs: UnisonStream + TypedFrame (handler-facing stream 型)
- dispatch.rs: handle_connection / client_accept_bi_loop (接続ディスパッチ)
- conn_quinn.rs: quinn 型 → UnisonConn/UnisonSend/UnisonRecv adapter impl
- quic.rs: QuicClient + QuicServer の raw-QUIC endpoint pair に集約

public API は不変。 frame / stream の項目は quic.rs から re-export し、
network::quic::* の import パスを維持。 logic 変更ゼロ、 全テスト green。
