+++ crates/unison-protocol/src/network/quic.rs
@@ -1,29 +1,34 @@
+//! Raw QUIC transport — [`QuicClient`] / [`QuicServer`] と接続ハンドラー。
+//!
+//! typed-frame の wire I/O は [`super::frame`]、 handler-facing なストリーム型
+//! [`UnisonStream`] は [`super::stream`] に分離されている。 後方互換のため、
+//! それらは本モジュールからも re-export されている (= `network::quic::*` の
+//! import パスを維持)。
+
 use anyhow::{Context, Result};
-use quinn::{ClientConfig, Connection, Endpoint, RecvStream, SendStream, ServerConfig};
+use quinn::{ClientConfig, Connection, Endpoint, ServerConfig};
 use rustls::pki_types::{CertificateDer, PrivateKeyDer};
 use rustls::{ClientConfig as RustlsClientConfig, ServerConfig as RustlsServerConfig};
 use std::net::SocketAddr;
-use std::pin::Pin;
-use std::sync::{
-    Arc,
-    atomic::{AtomicBool, AtomicU64, Ordering},
-};
-use tokio::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
+use std::sync::Arc;
 use tokio::sync::{Mutex, RwLock, mpsc, oneshot};
-use tracing::{debug, error, info, warn};
+use tracing::{error, info, warn};
 
-use super::conn::{BiStream, BoxUnisonRecv, BoxUnisonSend, UnisonConn, UnisonRecv, UnisonSend};
-use super::{
-    NetworkError, ProtocolFrame, ProtocolMessage, context::ConnectionContext,
-    server::ProtocolServer,
+use super::conn::UnisonConn;
+use super::dispatch::{client_accept_bi_loop, handle_connection};
+use super::{ProtocolMessage, context::ConnectionContext, server::ProtocolServer};
+
+// 後方互換: typed-frame wire I/O と handler-facing stream 型は専用モジュールへ
+// 移動したが、 `network::quic::*` の public import パスを保つため再公開する。
+pub use super::frame::{
+    CHANNEL_ACK_METHOD, FRAME_TYPE_PROTOCOL, FRAME_TYPE_RAW, read_frame, read_typed_frame,
+    write_frame, write_typed_frame,
 };
+pub use super::stream::{TypedFrame, UnisonStream};
 
 /// Default certificate file paths for assets/certs directory
 pub const DEFAULT_CERT_PATH: &str = "assets/certs/cert.pem";
 pub const DEFAULT_KEY_PATH: &str = "assets/certs/private_key.der";
-
-/// Maximum message size for QUIC streams (8MB)
-const MAX_MESSAGE_SIZE: usize = 8 * 1024 * 1024;
 
 /// Default port for QUIC connections
 const DEFAULT_PORT: u16 = 8080;
@@ -122,111 +127,6 @@ fn has_port(addr: &str) -> bool {
         }
     }
     false
-}
-
-/// Length-prefixed フレームの読み取り（4バイトBE長 + データ）
-/// ストリームを消費せずに1フレームだけ読む
-///
-/// `quinn::RecvStream` のみならず WebTransport のストリームでも使えるよう、
-/// `AsyncRead` でジェネリック化されている (= transport 非依存)。
-pub async fn read_frame<R: AsyncRead + Unpin + ?Sized>(recv: &mut R) -> Result<bytes::Bytes> {
-    let mut len_buf = [0u8; 4];
-    recv.read_exact(&mut len_buf)
-        .await
-        .context("Failed to read frame length")?;
-    let len = u32::from_be_bytes(len_buf) as usize;
-    if len > MAX_MESSAGE_SIZE {
-        return Err(anyhow::anyhow!("Frame too large: {} bytes", len));
-    }
-    let mut data = vec![0u8; len];
-    recv.read_exact(&mut data)
-        .await
-        .context("Failed to read frame data")?;
-    Ok(bytes::Bytes::from(data))
-}
-
-/// Length-prefixed フレームの書き込み
-pub async fn write_frame<W: AsyncWrite + Unpin + ?Sized>(send: &mut W, data: &[u8]) -> Result<()> {
-    let len = (data.len() as u32).to_be_bytes();
-    send.write_all(&len)
-        .await
-        .context("Failed to write frame length")?;
-    send.write_all(data)
-        .await
-        .context("Failed to write frame data")?;
-    Ok(())
-}
-
-/// フレームタイプタグ
-pub const FRAME_TYPE_PROTOCOL: u8 = 0x00;
-pub const FRAME_TYPE_RAW: u8 = 0x01;
-
-/// Channel open ack の method 名 (= Phase 6c)。
-///
-/// クライアントが `__channel:{name}` open frame を送ると、 サーバーは登録済み
-/// channel を lookup し、 同 stream へこの method の `ProtocolMessage` を 1 本返す:
-/// - **accept**: `msg_type = Response`、 `id` = open request の id、 payload `{}`
-/// - **nack** (= channel-not-found): `msg_type = Error`、 同 `id`、
-///   payload `{"error":"channel-not-found","channel":"{name}"}`
-///
-/// `id` が open request と一致するため、 クライアントは自分の open request に
-/// 相関させられる。 `__identity` と同じ `__`-prefix の特殊 method であり、 新しい
-/// typed frame type は追加しない (= 既存 wire layout は不変、 additive)。
-pub const CHANNEL_ACK_METHOD: &str = "__channel_ack";
-
-/// Typed フレーム — type tag 付きの読み書き
-/// フォーマット: [4 bytes: length][1 byte: type tag][payload]
-/// length は type tag + payload の合計バイト数
-///
-/// Typed フレームの読み取り — type tag とペイロードを返す
-pub async fn read_typed_frame<R: AsyncRead + Unpin + ?Sized>(
-    recv: &mut R,
-) -> Result<(u8, bytes::Bytes)> {
-    let mut len_buf = [0u8; 4];
-    recv.read_exact(&mut len_buf)
-        .await
-        .context("Failed to read frame length")?;
-    let total_len = u32::from_be_bytes(len_buf) as usize;
-    if total_len == 0 {
-        return Err(anyhow::anyhow!("Empty frame"));
-    }
-    if total_len > MAX_MESSAGE_SIZE {
-        return Err(anyhow::anyhow!("Frame too large: {} bytes", total_len));
-    }
-
-    // type tag を読む
-    let mut type_buf = [0u8; 1];
-    recv.read_exact(&mut type_buf)
-        .await
-        .context("Failed to read frame type tag")?;
-    let frame_type = type_buf[0];
-
-    // payload を読む
-    let payload_len = total_len - 1;
-    let mut data = vec![0u8; payload_len];
-    recv.read_exact(&mut data)
-        .await
-        .context("Failed to read frame payload")?;
-    Ok((frame_type, bytes::Bytes::from(data)))
-}
-
-/// Typed フレームの書き込み
-pub async fn write_typed_frame<W: AsyncWrite + Unpin + ?Sized>(
-    send: &mut W,
-    frame_type: u8,
-    data: &[u8],
-) -> Result<()> {
-    let total_len = (1 + data.len()) as u32;
-    send.write_all(&total_len.to_be_bytes())
-        .await
-        .context("Failed to write frame length")?;
-    send.write_all(&[frame_type])
-        .await
-        .context("Failed to write frame type tag")?;
-    send.write_all(data)
-        .await
-        .context("Failed to write frame payload")?;
-    Ok(())
 }
 
 /// QUIC client implementation
@@ -767,409 +667,6 @@ impl QuicServer {
     }
 }
 
-/// クライアント側: サーバー発信の双方向ストリームを受け付けるループ
-///
-/// サーバーが `connection.open_bi()` で開いたストリーム（Identity 送信等）を
-/// `accept_bi()` で受信し、ProtocolMessage に変換する。
-/// `__identity` メッセージは専用の oneshot チャネルに送り、それ以外は既存の mpsc に送る。
-async fn client_accept_bi_loop(
-    connection: Connection,
-    tx: mpsc::UnboundedSender<ProtocolMessage>,
-    identity_tx: Arc<Mutex<Option<oneshot::Sender<ProtocolMessage>>>>,
-) {
-    loop {
-        match connection.accept_bi().await {
-            Ok((_send_stream, mut recv_stream)) => {
-                let tx = tx.clone();
-                let identity_tx = identity_tx.clone();
-                tokio::spawn(async move {
-                    match read_typed_frame(&mut recv_stream).await {
-                        Ok((FRAME_TYPE_PROTOCOL, frame_bytes)) => {
-                            if let Ok(frame) = ProtocolFrame::from_bytes(&frame_bytes)
-                                && let Ok(message) = ProtocolMessage::from_frame(&frame)
-                            {
-                                if message.method == "__identity" {
-                                    // Identity メッセージは専用 oneshot チャネルに送信
-                                    if let Some(id_tx) = identity_tx.lock().await.take() {
-                                        let _ = id_tx.send(message);
-                                    } else {
-                                        warn!(
-                                            "Identity oneshot already consumed, dropping identity message"
-                                        );
-                                    }
-                                } else {
-                                    // それ以外は既存の mpsc チャネルに送信
-                                    let _ = tx.send(message);
-                                }
-                            }
-                        }
-                        Ok((frame_type, _)) => {
-                            warn!(
-                                "Unexpected frame type in server-initiated stream: 0x{:02x}",
-                                frame_type
-                            );
-                        }
-                        Err(e) => {
-                            warn!("Failed to read server-initiated stream: {}", e);
-                        }
-                    }
-                });
-            }
-            Err(quinn::ConnectionError::ApplicationClosed(_)) => {
-                info!("Connection closed by server");
-                break;
-            }
-            Err(e) => {
-                warn!("Failed to accept server-initiated stream: {}", e);
-                break;
-            }
-        }
-    }
-}
-
-/// Channel open ack / nack を 1 本の typed protocol frame として送信する (= Phase 6c)。
-///
-/// `accepted == true` なら [`MessageType::Response`] の `open_ack`、 `false` なら
-/// [`MessageType::Error`] の nack (= payload に `channel-not-found`) を `send`
-/// ストリームへ書き出す。 `request_id` は open request の id を引き継ぎ、
-/// クライアントが自分の open と相関できるようにする。
-async fn write_channel_ack<W: AsyncWrite + Unpin + ?Sized>(
-    send: &mut W,
-    request_id: u64,
-    accepted: bool,
-    channel_name: &str,
-) -> Result<()> {
-    use super::MessageType;
-
-    let (msg_type, payload) = if accepted {
-        (MessageType::Response, serde_json::json!({}))
-    } else {
-        (
-            MessageType::Error,
-            serde_json::json!({
-                "error": "channel-not-found",
-                "channel": channel_name,
-            }),
-        )
-    };
-    let msg = ProtocolMessage::new_with_json(
-        request_id,
-        CHANNEL_ACK_METHOD.to_string(),
-        msg_type,
-        payload,
-    )
-    .map_err(|e| anyhow::anyhow!("Failed to build open_ack message: {}", e))?;
-    let frame = msg
-        .into_frame()
-        .map_err(|e| anyhow::anyhow!("Failed to encode open_ack frame: {}", e))?;
-    write_typed_frame(send, FRAME_TYPE_PROTOCOL, &frame.to_bytes()).await
-}
-
-/// transport 非依存の接続ハンドラー。
-///
-/// raw QUIC と WebTransport の両 ingress がこの関数へ収束する。 `connection` は
-/// [`UnisonConn`] の trait object であり、 この関数は transport の種類を知らない。
-pub(crate) async fn handle_connection(
-    connection: Arc<dyn UnisonConn>,
-    server: Arc<ProtocolServer>,
-    ctx: Arc<ConnectionContext>,
-) -> Result<()> {
-    let remote_addr = connection.remote_address();
-
-    // v0.10.0: active connection に登録 (= server.broadcast の配信先)
-    let connection_arc = Arc::clone(&connection);
-    server
-        .add_active_connection(remote_addr, Arc::clone(&connection_arc))
-        .await;
-
-    // v0.10.0: datagram dispatcher を 1 connection に 1 個 spawn
-    // 登録された datagram channel handler 全てに対し、 channel_id を register して
-    // DatagramChannel を構築、 handler を別 task で起動
-    let datagram_handlers = server.snapshot_datagram_handlers().await;
-    let _datagram_dispatcher = if datagram_handlers.is_empty() {
-        // datagram handler が無ければ dispatcher を spawn しない (= overhead 回避)
-        None
-    } else {
-        let dispatcher = Arc::new(super::datagram_dispatcher::DatagramDispatcher::spawn(
-            Arc::clone(&connection_arc),
-        ));
-        for (name, channel_id, handler) in datagram_handlers {
-            let rx = dispatcher.register(channel_id, 256).await;
-            let datagram_channel = super::datagram_channel::DatagramChannel::<
-                crate::codec::JsonCodec,
-            >::new(
-                Arc::clone(&connection_arc), channel_id, name.clone(), rx
-            );
-            tokio::spawn(async move {
-                handler(datagram_channel).await;
-            });
-        }
-        Some(dispatcher)
-    };
-
-    // Identity Handshake: 接続直後にServerIdentityを送信
-    let identity = server.build_identity().await;
-    ctx.set_identity(identity.clone()).await;
-
-    let identity_msg = identity.to_protocol_message();
-    match identity_msg.into_frame() {
-        Ok(frame) => {
-            let frame_bytes = frame.to_bytes();
-            match connection.open_bi().await {
-                Ok((mut send_stream, _recv_stream)) => {
-                    if let Err(e) =
-                        write_typed_frame(&mut send_stream, FRAME_TYPE_PROTOCOL, &frame_bytes).await
-                    {
-                        warn!("Failed to send identity: {}", e);
-                    } else {
-                        let _ = send_stream.finish().await;
-                        info!("Identity sent to client");
-                    }
-                }
-                Err(e) => {
-                    warn!("Failed to open identity stream: {}", e);
-                }
-            }
-            // 注: WebTransport セッションにも同一フローが適用される。
-        }
-        Err(e) => {
-            warn!("Failed to serialize identity frame: {}", e);
-        }
-    }
-
-    // 接続イベントを送信
-    server.emit_connection_event(super::server::ConnectionEvent::Connected {
-        remote_addr,
-        context: Arc::clone(&ctx),
-    });
-
-    loop {
-        let connection_clone = Arc::clone(&connection);
-        match connection.accept_bi().await {
-            Ok((send_stream, mut recv_stream)) => {
-                let server = Arc::clone(&server);
-                let connection = connection_clone;
-                let ctx = Arc::clone(&ctx);
-
-                tokio::spawn(async move {
-                    // typed frame で読み取り（type tag 付き）
-                    let request_result = match read_typed_frame(&mut recv_stream).await {
-                        Ok((FRAME_TYPE_PROTOCOL, frame_bytes)) => {
-                            ProtocolFrame::from_bytes(&frame_bytes)
-                                .and_then(|frame| ProtocolMessage::from_frame(&frame))
-                        }
-                        Ok((frame_type, _)) => {
-                            warn!("Unexpected frame type in handshake: 0x{:02x}", frame_type);
-                            return;
-                        }
-                        Err(e) => {
-                            error!("Failed to read handshake frame: {}", e);
-                            return;
-                        }
-                    };
-
-                    match request_result {
-                        Ok(request) => {
-                            // チャネルルーティング: __channel: プレフィックスをチェック
-                            if let Some(channel_name) = request.method.strip_prefix("__channel:") {
-                                let channel_name = channel_name.to_string();
-                                let mut send_stream = send_stream;
-                                if let Some(handler) =
-                                    server.get_channel_handler(&channel_name).await
-                                {
-                                    // channel lifecycle の "open" 側ログ。
-                                    // close 側 (= 下記の debug!) と対になり、 1 接続中の
-                                    // channel 開閉 trace が debug level で揃う。
-                                    // info level にしない理由: 1 接続で channel が頻繁に
-                                    // open/close される設計 (= 1 request/response = 1 channel)
-                                    // なので info noise になりがち。
-                                    debug!("Channel '{}' opened", channel_name);
-
-                                    // Phase 6c: open frame と同 stream へ open_ack
-                                    // (= Response) を 1 本返す。 id は open request の
-                                    // id を引き継ぎ、 クライアントが相関できるようにする。
-                                    if let Err(e) = write_channel_ack(
-                                        &mut send_stream,
-                                        request.id,
-                                        true,
-                                        &channel_name,
-                                    )
-                                    .await
-                                    {
-                                        warn!(
-                                            "Failed to send open_ack for '{}': {}",
-                                            channel_name, e
-                                        );
-                                        return;
-                                    }
-
-                                    // チャネル用のUnisonStreamを作成（ストリームは生きたまま）
-                                    let stream = UnisonStream::from_streams(
-                                        request.id,
-                                        request.method.clone(),
-                                        connection,
-                                        send_stream,
-                                        recv_stream,
-                                    );
-                                    if let Err(e) = handler(ctx, stream).await {
-                                        // sender 側が request/response 完了後に正常 close した
-                                        // end-of-stream は real error ではないので debug level に
-                                        // degrade。 これにより毎 channel session の終端で発生する
-                                        // ERROR log noise (= journal で大半を占める) を抑制。
-                                        if e.is_normal_close() {
-                                            debug!(
-                                                "Channel '{}' closed normally (end of stream)",
-                                                channel_name
-                                            );
-                                        } else {
-                                            error!(
-                                                "Channel handler error for '{}': {}",
-                                                channel_name, e
-                                            );
-                                        }
-                                    }
-                                } else {
-                                    // Phase 6c: 未登録 channel への open は nack
-                                    // (= Error frame) を返してから stream を畳む。
-                                    // これによりクライアントの open は silent に
-                                    // hang せず channel-not-found で即 reject する。
-                                    warn!("No channel handler for: {}", channel_name);
-                                    if let Err(e) = write_channel_ack(
-                                        &mut send_stream,
-                                        request.id,
-                                        false,
-                                        &channel_name,
-                                    )
-                                    .await
-                                    {
-                                        warn!(
-                                            "Failed to send open nack for '{}': {}",
-                                            channel_name, e
-                                        );
-                                    } else {
-                                        let _ = send_stream.finish().await;
-                                    }
-                                }
-                                return;
-                            }
-
-                            // 非チャネルメッセージはサポート外
-                            warn!(
-                                "Non-channel message received (method: {}). Use channels instead.",
-                                request.method
-                            );
-                        }
-                        Err(e) => {
-                            warn!("Failed to parse message: {}", e);
-                        }
-                    }
-                });
-            }
-            Err(e) => {
-                // accept_bi の Err = 接続終了 (= 正常な切断もエラー扱いで来る)。
-                // transport を問わず接続ループを抜ける。
-                info!("Connection closed ({}), client disconnected", e);
-                server.emit_connection_event(super::server::ConnectionEvent::Disconnected {
-                    remote_addr,
-                });
-                break;
-            }
-        }
-    }
-
-    // v0.10.0: connection 終了時に active_connections から remove
-    // (= broadcast 配信先から自動除外、 datagram dispatcher は _datagram_dispatcher 変数の
-    // scope-exit drop で同時に abort される)
-    server.remove_active_connection(remote_addr).await;
-
-    Ok(())
-}
-
-// ─────────────────────────────────────────
-// UnisonConn 実装 — quinn::Connection
-// ─────────────────────────────────────────
-
-/// `quinn::SendStream` を [`UnisonSend`] として扱う。
-///
-/// `quinn::SendStream` は `tokio::io::AsyncWrite` を実装済みなので、 `finish`
-/// だけ橋渡しすればよい。
-impl UnisonSend for SendStream {
-    fn finish(
-        &mut self,
-    ) -> Pin<Box<dyn std::future::Future<Output = Result<(), NetworkError>> + Send + '_>> {
-        // quinn の finish は同期。 trait の async シグネチャに合わせて即解決する
-        // future で包む。 二重 finish (= ClosedStream) は正常終了扱い (= 冪等)。
-        let result = match SendStream::finish(self) {
-            Ok(()) | Err(quinn::ClosedStream { .. }) => Ok(()),
-        };
-        Box::pin(async move { result })
-    }
-}
-
-/// `quinn::RecvStream` を [`UnisonRecv`] として扱う。
-impl UnisonRecv for RecvStream {
-    fn stop(&mut self) -> Result<(), NetworkError> {
-        match RecvStream::stop(self, quinn::VarInt::from_u32(0)) {
-            Ok(()) | Err(quinn::ClosedStream { .. }) => Ok(()),
-        }
-    }
-}
-
-/// `quinn::Connection` を transport-agnostic な [`UnisonConn`] として公開する。
-impl UnisonConn for Connection {
-    fn accept_bi(
-        &self,
-    ) -> Pin<Box<dyn std::future::Future<Output = Result<BiStream, NetworkError>> + Send + '_>>
-    {
-        Box::pin(async move {
-            let (send, recv) = Connection::accept_bi(self)
-                .await
-                .map_err(|e| NetworkError::Quic(format!("accept_bi failed: {}", e)))?;
-            let send: BoxUnisonSend = Box::new(send);
-            let recv: BoxUnisonRecv = Box::new(recv);
-            Ok((send, recv))
-        })
-    }
-
-    fn open_bi(
-        &self,
-    ) -> Pin<Box<dyn std::future::Future<Output = Result<BiStream, NetworkError>> + Send + '_>>
-    {
-        Box::pin(async move {
-            let (send, recv) = Connection::open_bi(self)
-                .await
-                .map_err(|e| NetworkError::Quic(format!("open_bi failed: {}", e)))?;
-            let send: BoxUnisonSend = Box::new(send);
-            let recv: BoxUnisonRecv = Box::new(recv);
-            Ok((send, recv))
-        })
-    }
-
-    fn send_datagram(&self, data: bytes::Bytes) -> Result<(), NetworkError> {
-        Connection::send_datagram(self, data)
-            .map_err(|e| NetworkError::Quic(format!("send_datagram failed: {}", e)))
-    }
-
-    fn recv_datagram(
-        &self,
-    ) -> Pin<Box<dyn std::future::Future<Output = Result<bytes::Bytes, NetworkError>> + Send + '_>>
-    {
-        Box::pin(async move {
-            Connection::read_datagram(self)
-                .await
-                .map_err(|e| NetworkError::Quic(format!("recv_datagram failed: {}", e)))
-        })
-    }
-
-    fn remote_address(&self) -> SocketAddr {
-        Connection::remote_address(self)
-    }
-
-    fn close(&self, code: u32, reason: &[u8]) {
-        Connection::close(self, quinn::VarInt::from_u32(code), reason);
-    }
-}
-
 /// Server-side cert resolver that always returns the same [`rustls::sign::CertifiedKey`].
 ///
 /// Holds the key behind a single `Arc` so the private key material exists in
@@ -1183,196 +680,6 @@ impl rustls::server::ResolvesServerCert for SingleCertResolver {
         _client_hello: rustls::server::ClientHello<'_>,
     ) -> Option<Arc<rustls::sign::CertifiedKey>> {
         Some(Arc::clone(&self.0))
-    }
-}
-
-/// Unison Stream — transport 非依存の双方向ストリーム実装。
-///
-/// 内部は [`BoxUnisonSend`] / [`BoxUnisonRecv`] (= trait object) を保持し、
-/// raw QUIC と WebTransport のどちらのストリームでも同一の型で扱える。 これが
-/// `register_channel` ハンドラーに渡る面 (= handler-facing API) なので、 型名・
-/// メソッドシグネチャは安定させている。
-pub struct UnisonStream {
-    stream_id: u64,
-    method: String,
-    #[allow(dead_code)]
-    connection: Arc<dyn UnisonConn>,
-    send_stream: Arc<Mutex<Option<BoxUnisonSend>>>,
-    recv_stream: Arc<Mutex<Option<BoxUnisonRecv>>>,
-    is_active: Arc<AtomicBool>,
-}
-
-impl UnisonStream {
-    pub async fn new(
-        method: String,
-        connection: Arc<dyn UnisonConn>,
-        stream_id: Option<u64>,
-    ) -> Result<Self> {
-        static STREAM_ID_COUNTER: AtomicU64 = AtomicU64::new(1);
-
-        let id = stream_id.unwrap_or_else(|| STREAM_ID_COUNTER.fetch_add(1, Ordering::SeqCst));
-
-        // Open bidirectional stream
-        let (send_stream, recv_stream) = connection
-            .open_bi()
-            .await
-            .context("Failed to open bidirectional stream")?;
-
-        Ok(Self {
-            stream_id: id,
-            method,
-            connection,
-            send_stream: Arc::new(Mutex::new(Some(send_stream))),
-            recv_stream: Arc::new(Mutex::new(Some(recv_stream))),
-            is_active: Arc::new(AtomicBool::new(true)),
-        })
-    }
-
-    /// 既存のストリームから作成（サーバー側）
-    pub fn from_streams(
-        stream_id: u64,
-        method: String,
-        connection: Arc<dyn UnisonConn>,
-        send_stream: BoxUnisonSend,
-        recv_stream: BoxUnisonRecv,
-    ) -> Self {
-        Self {
-            stream_id,
-            method,
-            connection,
-            send_stream: Arc::new(Mutex::new(Some(send_stream))),
-            recv_stream: Arc::new(Mutex::new(Some(recv_stream))),
-            is_active: Arc::new(AtomicBool::new(true)),
-        }
-    }
-
-    /// ストリーム稼働状態の確認
-    pub fn is_active(&self) -> bool {
-        self.is_active.load(Ordering::SeqCst)
-    }
-}
-
-/// Typed フレーム受信結果
-pub enum TypedFrame {
-    /// ProtocolMessage フレーム (type tag 0x00)
-    Protocol(ProtocolMessage),
-    /// Raw bytes フレーム (type tag 0x01)
-    Raw(Vec<u8>),
-}
-
-impl UnisonStream {
-    /// ProtocolMessage を typed フレームとして送信（type tag 0x00）
-    ///
-    /// SystemStream::send() を経由せず、ProtocolMessage → into_frame() → write_typed_frame() で
-    /// type tag 付き length-prefixed フレームとして送信する。チャネル通信で使用。
-    pub async fn send_frame(&self, msg: &ProtocolMessage) -> Result<(), NetworkError> {
-        if !self.is_active() {
-            return Err(NetworkError::Connection("Stream is not active".to_string()));
-        }
-
-        let frame = msg.clone().into_frame()?;
-        let frame_bytes = frame.to_bytes();
-
-        let mut send_guard = self.send_stream.lock().await;
-        if let Some(send_stream) = send_guard.as_mut() {
-            write_typed_frame(send_stream, FRAME_TYPE_PROTOCOL, &frame_bytes)
-                .await
-                .map_err(|e| NetworkError::Quic(format!("Failed to send frame: {}", e)))?;
-            Ok(())
-        } else {
-            Err(NetworkError::Connection(
-                "Send stream is closed".to_string(),
-            ))
-        }
-    }
-
-    /// Raw bytes を typed フレームとして送信（type tag 0x01）
-    ///
-    /// buffa/zstd をバイパスし、length-prefix + type tag + raw payload のみ。
-    /// オーディオストリーミング等の最小オーバーヘッド通信に使用。
-    pub async fn send_raw_frame(&self, data: &[u8]) -> Result<(), NetworkError> {
-        if !self.is_active() {
-            return Err(NetworkError::Connection("Stream is not active".to_string()));
-        }
-
-        let mut send_guard = self.send_stream.lock().await;
-        if let Some(send_stream) = send_guard.as_mut() {
-            write_typed_frame(send_stream, FRAME_TYPE_RAW, data)
-                .await
-                .map_err(|e| NetworkError::Quic(format!("Failed to send raw frame: {}", e)))?;
-            Ok(())
-        } else {
-            Err(NetworkError::Connection(
-                "Send stream is closed".to_string(),
-            ))
-        }
-    }
-
-    /// ストリームを閉じる（&self で呼べるバージョン、Arc 共有時に使用）
-    pub async fn close_stream(&self) -> Result<(), NetworkError> {
-        self.is_active.store(false, Ordering::SeqCst);
-
-        if let Some(mut send_stream) = self.send_stream.lock().await.take() {
-            send_stream.finish().await?;
-        }
-
-        if let Some(mut recv_stream) = self.recv_stream.lock().await.take() {
-            recv_stream.stop()?;
-        }
-
-        info!(
-            "Stream {} closed for method '{}'",
-            self.stream_id, self.method
-        );
-        Ok(())
-    }
-
-    /// ProtocolMessage のみを受信（後方互換）
-    ///
-    /// typed frame を読んで ProtocolMessage のみを返す。
-    /// Raw bytes フレームが来た場合はエラーを返す。
-    pub async fn recv_frame(&self) -> Result<ProtocolMessage, NetworkError> {
-        match self.recv_typed_frame().await? {
-            TypedFrame::Protocol(msg) => Ok(msg),
-            TypedFrame::Raw(_) => Err(NetworkError::Protocol(
-                "Expected protocol frame, got raw bytes".to_string(),
-            )),
-        }
-    }
-
-    /// Typed フレームを受信（ProtocolMessage or Raw bytes）
-    ///
-    /// type tag で振り分けて TypedFrame を返す。
-    /// チャネルの recv ループで使用し、Protocol/Raw を適切なキューに振り分ける。
-    pub async fn recv_typed_frame(&self) -> Result<TypedFrame, NetworkError> {
-        if !self.is_active() {
-            return Err(NetworkError::Connection("Stream is not active".to_string()));
-        }
-
-        let mut recv_guard = self.recv_stream.lock().await;
-        if let Some(recv_stream) = recv_guard.as_mut() {
-            let (frame_type, payload) = read_typed_frame(recv_stream).await.map_err(|e| {
-                self.is_active.store(false, Ordering::SeqCst);
-                NetworkError::Quic(format!("Failed to read frame: {}", e))
-            })?;
-
-            match frame_type {
-                FRAME_TYPE_PROTOCOL => {
-                    let frame = ProtocolFrame::from_bytes(&payload)?;
-                    let message = ProtocolMessage::from_frame(&frame)?;
-                    Ok(TypedFrame::Protocol(message))
-                }
-                FRAME_TYPE_RAW => Ok(TypedFrame::Raw(payload.to_vec())),
-                _ => Err(NetworkError::Protocol(format!(
-                    "Unknown frame type tag: 0x{:02x}",
-                    frame_type
-                ))),
-            }
-        } else {
-            Err(NetworkError::Connection(
-                "Receive stream is closed".to_string(),
-            ))
-        }
     }
 }
